main_app/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
langgraph>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
google-generativeai>=0.7.0
pydantic>=2.0.0
supabase>=2.13.0
pinecone>=7.3.0
langchain-pinecone>=0.2.13
pypdf>=6.9.2
